refactor: validator.js 파일 이동
- utils 폴더에서 sevice 폴더로 이동
